requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
pandas>=2.1.0
numpy>=1.24.0
python-binance>=1.0.19
ccxt>=4.0.0
pyyaml>=6.0
kiteconnect>=4.0.0
smartapi-python>=1.3.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.13.0
transformers>=4.30.0
yfinance>=0.2.0
nltk>=3.8.0
spacy>=3.6.0
plotly>=5.17.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
